Update more files in mach-msm with newer code from CAF
mach-msm in 3.13 has considerably older code than is present in msm-3.10
CAF repo. If we're going to keep things as up to date as possible, we're
going to want to use CAF's sources as a base for mach-msm; and linux
mainline as a base for everything else.

Fortunately, 3.10 is not far behind in terms of version number and
there's not much between the two versions that breaks compatibility,
making this easy in most cases.