Update cpufreq.c/h; Add CPU_FREQ/CPU_FREQ_MSM to config
Update to the latest MSM cpufreq driver from CAF (as of the end of
January). Because we're turning on CPUFREQ for the first time, we need
ARCH_APQ8064 to select ARCH_HAS_CPUFREQ to tell CPUFREQ we have support.

For now, we'll set our default governor to Ondemand with
Performance/Powersave/Userspace/Conservative as options. In the future,
this will probably be set to Faux123's Intelliactive.

Update acpuclock-krait to avoid breakage with the new CPUFREQ drivers.

Update include/trace/events/power.h to support
cpu_frequency_switch_start & cpu_frequency_switch_end.