arm/Kconfig: update
Stop selecting COMMON_CLK. It's causing redefinitions and it's not
selected by CAF's Kconfig.

The "Qualcomm MSM" section of this Kconfig is just as outdated as the
files in mach-msm from upstream. Select HAVE_CLK and HAVE_CLK_PREPARE
for now and possibly more using MSM-3.10's Kconfig as a source later.

Update clkdev.h to include_of_idx for the qcom clk drivers.