TEST: add smp support to board-jf_tmo
Fortunately, the platsmp.c imported from MSM-3.10 includes support for
the MSM8960. As the MSM8960 and APQ8064 share a lot in common, this
should work and enable SMP support for the APQ8064.

It is possible that some older drivers that use memory barriers will
need to be changed to use the SMP version of the functions.