mach-msm: remove MSM_VIC; add USE_OF and DEBUG_FS to Makefile
CONFIG_MSM_VIC is not used in any defconfig or kernel as far as I can
tell. It doesn't exist in newer MSM kernels either, so we can probably
safely assume that it's not needed.

Add CONFIG_USE_OF and CONFIG_DEBUG_FS so that our Makefile more
consistently resembles the Makefile our sources are (mostly) based on,
in case they are needed.