MDP: fix regulator_enable warnings and add some micro-optimizations
These lines don't have else conditionals afterwards and it appears as
though they should always be on, so we can micro-optimize things to tell
the branch predictor that it's likely dsi_pll_vdda and dsi_pll_vddio
exist.

Likewise, it's unlikely that the return value of the first two
regulator_enables will be a failure, so when we check for the return
values we can use unlikely for them, but not for ret3 because it's
possible it can go either way.